modification base donee

+++ backend/DTOs/ConsommationCapexDto.cs
@@ -5,20 +5,20 @@ using System.Text.Json.Serialization;
 public class ConsommationDepartementDto
 {
     public string DepartementNom { get; set; } = string.Empty;
-    public decimal MontantConsomme { get; set; }
+    public double MontantConsomme { get; set; }
 }
 
 public class ConsommationCapexDto
 {
     public int CapexId { get; set; }
     public string NomCapex { get; set; } = string.Empty;
-    public decimal BudgetTotal { get; set; }
-    public decimal BudgetRestant { get; set; }
+    public double BudgetTotal { get; set; }
+    public double BudgetRestant { get; set; }
     [JsonIgnore]
-    public decimal ResteBudget { get => BudgetRestant; set => BudgetRestant = value; }
-    public decimal BudgetRestantStocke { get; set; }
-    public decimal BudgetRestantCalcule { get; set; }
+    public double ResteBudget { get => BudgetRestant; set => BudgetRestant = value; }
+    public double BudgetRestantStocke { get; set; }
+    public double BudgetRestantCalcule { get; set; }
     public bool ResteBudgetIncoherent { get; set; }
-    public decimal MontantEnAttente { get; set; }
+    public double MontantEnAttente { get; set; }
     public List<ConsommationDepartementDto> ParDepartement { get; set; } = new();
 }
