+++ backend/DTOs/CreateCapexDto.cs
@@ -3,5 +3,5 @@ namespace backend.DTOs;
 public class CreateCapexDto
 {
     public string NomCapex { get; set; } = string.Empty;
-    public decimal BudgetTotal { get; set; }
+    public double BudgetTotal { get; set; }
 }
