+++ backend/Data/Repositories/ICapexRepository.cs
@@ -6,8 +6,8 @@ public interface ICapexRepository
     Task<Capex?> GetByIdAsync(int id);
     Task<List<Capex>> GetAllAsync();
     Task<int> AddAsync(Capex capex);
-    Task<bool> DecrementerResteBudgetAsync(int capexId, decimal montant);
+    Task<bool> DecrementerResteBudgetAsync(int capexId, double montant);
     Task<List<ConsommationDepartementDto>> GetConsommationParDepartementAsync(int capexId);
 
-    Task<decimal> GetMontantEnAttenteAsync(int capexId);
+    Task<double> GetMontantEnAttenteAsync(int capexId);
     }
